requirements.txt
slack_sdk==3.21.0
python-dotenv==1.0.1
bs4
requests
schedule==1.2.2